MSVC project update for Animals/ etc and warnings/fixes

+++ Source/Palette.cpp
@@ -39,7 +39,8 @@ const Colour& Palette::GetColour(int key) const
 
 const Colour& Palette::GetColour(float f) const
 {
-  int key = std::round(f * static_cast<float>(m_colours.size() - 1));
+  int key = static_cast<int>(
+    std::round(f * static_cast<float>(m_colours.size() - 1)));
   key = std::clamp(key, 0, static_cast<int>(m_colours.size() - 1));
   return m_colours[key];
 }
